fix(新手引导): 新建用户-新手引导弹窗触发问题处理
