+++ project-management/frontend/src/business/home/ProjectHome.vue
@@ -170,13 +170,8 @@ export default {
   },
   mounted() {
     this.$refs.introduction.resVisible = false
-    let microApps = JSON.parse(sessionStorage.getItem("micro_apps"));
-    if(localStorage.getItem("guide") === 'false' && microApps && microApps['project']) {
-      if(localStorage.getItem("step") === '3'){
-        this.initStep()
-      }else{
-        this.initStepAll()
-      }
+    if((localStorage.getItem("guide") === 'false' || localStorage.getItem("guide") === 'click') && localStorage.getItem("step") === '3') {
+      this.initStep()
     }
   },
   methods: {
@@ -191,151 +186,6 @@ export default {
     },
     edit() {
       this.$refs.editProject.edit(this.project);
-    },
-    initStepAll() {
-      const _this = this
-      _this.$nextTick(() => {
-        const tour = _this.$shepherd({
-          useModalOverlay: true,
-          exitOnEsc: false,
-          keyboardNavigation: false,
-          defaultStepOptions: {
-            scrollTo: {
-              behavior: 'smooth',
-              block: 'center'
-            },
-            canClickTarget: false,
-            // 高亮元素四周要填充的空白像素
-            modalOverlayOpeningPadding: 0,
-            // 空白像素的圆角
-            modalOverlayOpeningRadius: 4
-          }
-        })
-        tour.addSteps([
-          {
-            attachTo: {
-              element: document.querySelector('.shepherd-workspace'),
-              on: 'bottom-start'
-            },
-            buttons: [
-              {
-                action: function() {
-                  _this.$refs.introduction.resVisible = true
-                  return _this.gotoCancel(this, true)
-                },
-                classes: 'close-btn',
-                text: _this.$t("shepherd.exit")
-              },
-              {
-                action: function() {
-                  return _this.gotoNext(this, null, 2)
-                },
-                classes: 'shep-btn',
-                text: _this.$t("shepherd.next")
-              }
-            ],
-            title: _this.$t("shepherd.step1.title"),
-            text: _this.$t("shepherd.step1.text")
-          },
-          {
-            attachTo: {
-              element: document.querySelector('.shepherd-menu'),
-              on: 'right'
-            },
-            buttons: [
-              {
-                action: function() {
-                  _this.$refs.introduction.resVisible = localStorage.getItem("step") > 1
-                  return _this.gotoCancel(this, true)
-                },
-                classes: 'close-btn',
-                text: _this.$t("shepherd.exit")
-              },
-              {
-                action: function() {
-                  return _this.gotoNext(this, '/project/home', 3)
-                },
-                classes: 'shep-btn',
-                text: _this.$t("shepherd.next")
-              }
-            ],
-            title: _this.$t("shepherd.step2.title"),
-            text: _this.$t("shepherd.step2.text")
-          },
-          {
-            attachTo: {
-              element: document.querySelector('.shepherd-project'),
-              on: 'bottom-end'
-            },
-            classes: "custom-width",
-            buttons: [
-              {
-                action: function() {
-                  _this.$refs.introduction.resVisible = localStorage.getItem("step") > 1
-                  return _this.gotoCancel(this, true)
-                },
-                classes: 'close-btn',
-                text: _this.$t("shepherd.exit")
-              },
-              {
-                action: function() {
-                  return _this.gotoNext(this, null, 4)
-                },
-                classes: 'shep-btn',
-                text: _this.$t("shepherd.next")
-              }
-            ],
-            title: _this.$t("shepherd.step3.title"),
-            text: _this.$t("shepherd.step3.text")
-          },
-          {
-            attachTo: {
-              element: document.querySelector('.shepherd-project-menu'),
-              on: 'bottom-start'
-            },
-            buttons: [
-              {
-                action: function() {
-                  _this.$refs.introduction.resVisible = localStorage.getItem("step") > 1
-                  return _this.gotoCancel(this, true)
-                },
-                classes: 'close-btn',
-                text: _this.$t("shepherd.exit")
-              },
-              {
-                action: function() {
-                  return _this.gotoNext(this, '/project/home', 5)
-                },
-                classes: 'shep-btn',
-                text: _this.$t("shepherd.next")
-              }
-            ],
-            title: _this.$t("shepherd.step4.title"),
-            text: _this.$t("shepherd.step4.text")
-          },
-          {
-            arrow:true,
-            modalOverlayOpeningPadding: 8,
-            attachTo: {
-              element: document.querySelector('.shepherd-project-name'),
-              on: 'right'
-            },
-            buttons: [
-              {
-                action: function() {
-                  _this.$refs.introduction.resVisible = localStorage.getItem("step") > 1
-                  return _this.gotoCancel(this, false)
-                },
-                classes: 'close-btn',
-                text: _this.$t("shepherd.know")
-              }
-            ],
-            title: _this.$t("shepherd.step5.title"),
-            text: _this.$t("shepherd.step5.text")
-          }
-        ])
-        tour.start()
-      })
     },
     initStep() {
       const _this = this
